% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phx_crime.R
\docType{data}
\name{phx_crime}
\alias{phx_crime}
\title{Phoenix Crime Incidents}
\format{
A `data.frame` object with 624,147 incidents.
}
\source{
Data retrieved from \url{https://www.phoenixopendata.com/dataset/ouof/resource/c79b2135-e936-439e-a8a3-79e61d4518d2}
}
\usage{
phx_crime
}
\description{
An object of class `data.frame`.
}
\details{
This dataset contains crime incidents from 2015-2025 reported by the Phoenix Police Department.
}
\examples{
View( phx_crime )
}
\keyword{datasets}
