% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phx_uof_2018_2024.R
\docType{data}
\name{phx_uof_2018_2024}
\alias{phx_uof_2018_2024}
\title{Phoenix Use of Force Incidents}
\format{
A `data.frame` object with 12,324 incidents and 57 varibles.
}
\source{
Data retrieved from \url{https://www.phoenixopendata.com/dataset/ouof/resource/c79b2135-e936-439e-a8a3-79e61d4518d2}
}
\usage{
phx_uof_2018_2024
}
\description{
An object of class `data.frame`.
}
\details{
This dataset contains Use of Force (UOF) incidents from 2018-2024 reported by the Phoenix Police Department.
}
\examples{
View( phx_uof_2018_2024 )
}
\keyword{datasets}
