% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempe-hate.R
\docType{data}
\name{tempe_hate}
\alias{tempe_hate}
\title{Tempe Hate Crime Incidents}
\format{
A `data.frame` object with 71 incidents.
}
\source{
Data retrieved from \url{https://data.tempe.gov/}
}
\usage{
tempe_hate
}
\description{
An object of class `data.frame`.
}
\details{
This dataset contains hate crime incidents reported to the Tempe Police Department.
}
\examples{
View( tempe_hate )
}
\keyword{datasets}
