% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troubles-fatalities.R
\docType{data}
\name{troubles_fatalities}
\alias{troubles_fatalities}
\title{Troubles Fatalities}
\format{
A `data.frame` object with 3,719 incidents.
}
\source{
Adida, C. L., Brown, J. M., McCord, G. C., & McLachlan, P. (2023).
Diversity and violence during conflict migration: The Troubles in Northern Ireland.
Political Science Research and Methods, 11(3), 451--467.
doi:10.1017/psrm.2022.36
}
\usage{
troubles_fatalities
}
\description{
An object of class `data.frame`.
}
\details{
This dataset contains fatalities occurring during the Troubles in Northern Ireland.
}
\examples{
View( troubles_fatalities )
}
\keyword{datasets}
