% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tucson-bike.R
\docType{data}
\name{tucson_bike}
\alias{tucson_bike}
\title{Tucson Bicycle Incidents}
\format{
A `data.frame` object with 1,845 incidents.
}
\source{
Data retrieved from \url{https://gisdata.tucsonaz.gov/}
}
\usage{
tucson_bike
}
\description{
An object of class `data.frame`.
}
\details{
This dataset contains traffic incidents involving a bicycle reported to the Tucson Police Department.
}
\examples{
View( tucson_bike )
}
\keyword{datasets}
