% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tucson-traffic.R
\docType{data}
\name{tucson_traffic}
\alias{tucson_traffic}
\title{Tucson Traffic Incidents}
\format{
A `data.frame` object with 53,547 incidents.
}
\source{
Data retrieved from \url{https://gisdata.tucsonaz.gov/}
}
\usage{
tucson_traffic
}
\description{
An object of class `data.frame`.
}
\details{
This dataset contains traffic incidents reported to the Tucson Police Department.
}
\examples{
View( tucson_traffic )
}
\keyword{datasets}
